+++ src/Todo.svelte
@@ -1,14 +1,11 @@
 <script>
   import { createEventDispatcher } from "svelte";
+  import todos from "./todos.js";
+
   export let todo;
-  const dispatch = createEventDispatcher();
 
   function handleChange() {
-    if (todo.done) {
-      dispatch("uncheck", todo.id);
-    } else {
-      dispatch("check", todo.id);
-    }
+    (todo.done ? todos.uncheck : todos.check)(todo.id);
   }
 </script>
 
@@ -21,7 +18,11 @@
 <li>
   <label>
     {#if todo.done}◉{:else}○{/if}
-    <span class:done="{todo.done}">{todo.text}</span>
-    <input type="checkbox" checked={todo.done} on:change|preventDefault="{handleChange}" style="display: none;" />
+    <span class:done={todo.done}>{todo.text}</span>
+    <input
+      type="checkbox"
+      checked={todo.done}
+      on:change|preventDefault={handleChange}
+      style="display: none;" />
   </label>
 </li>
