+++ src/TodoList.svelte
@@ -10,9 +10,9 @@
 {#if filteredTodos && filteredTodos.length}
   <ul>
     {#each filteredTodos as todo (todo.id)}
-      <Todo {todo} on:check on:uncheck />
+      <Todo {todo} />
     {/each}
   </ul>
 {:else}
-  <p>... no items to show ...</p>
+  <p>no items to show ...</p>
 {/if}
